objetivo 3 parcialmente completo. falta descobrir como rodar o jpa e dps o schema.sql

+++ src/main/java/com/bcopstein/ex4_lancheriaddd_v1/Dominio/Entidades/ItemEstoque.java
@@ -11,7 +11,7 @@ public class ItemEstoque {
         this.id = id;
     }
 
-    public int getId(){return this.id;}
+    public long getId(){return this.id;}
     public Ingrediente getIngrediente() { return ingrediente; }
     public int getQuantidade() { return quantidade; }
     public void setQuantidade(int quantidade) { this.quantidade = quantidade; }
